Moved integration test to new file
